requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
psycopg2-binary
SQLAlchemy
pypdf
docx2txt
numpy